pyproject.toml
[project]
name = "cloud-linux-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-guardian>=2.4.0",
    "django>=5.1.3",
    "psycopg[pool]>=3.2.3",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
]

[tool.uv]
dev-dependencies = [
    "django-stubs[compatible-mypy]>=5.1.1",
    "mypy>=1.13.0",
    "pytest-django>=4.9.0",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
]

[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.ruff.lint]
select = ["E", "F", "B", "I"]


[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
quote-style = 'preserve'

[tool.mypy]
strict = true
ignore_missing_imports = true
mypy_path = "./app"
plugins = [
    "mypy_django_plugin.main",
    "pydantic.mypy"
]

[tool.django-stubs]
django_settings_module = "example.settings"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "example.settings"
